requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
coreapi==2.3.3
coreschema==0.0.4
cycler==0.12.1
Django==4.2.13
django-bootstrap-v5==1.0.11
django-cors-headers==4.3.1
django-extra-fields==3.0.2
djangorestframework==3.15.1
fonttools==4.53.0
fpdf==1.7.2
idna==3.7
itypes==1.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mysqlclient==2.2.4
numpy==2.0.0
packaging==24.1
pillow==10.3.0
psycopg2==2.9.9
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.32.3
setuptools==70.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
